add maze_pre_gpu
